Reconciling ImguiLayer and ExampleLayer Vulkan adjustments - IV
Improved self notion of FRAMES_ON_FLIGHT and ImageIndex and made relevant changes in VulkanRendererAPI.
Refactored huge chunk of Dear ImGUI's code blocks. Now it should be more in alignment with Karma's philosophy. Need to address the concern raised by the author. For instance checking of the need for "separate viewport".